requirements.txt
fire
simple-parsing
pyyaml
mistral-common>=1.3.1
safetensors
tensorboard
tqdm

torch==2.2
triton==2.2
xformers==0.0.24

huggingface_hub
wandb
tqdm
pandas
pyarrow
fastparquet
ruamel.yaml